Move Kotlin parameter opmodes into the superclass' file
