+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/UnifiedTeleOpKt.kt
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.opModes
 import com.acmerobotics.dashboard.FtcDashboard
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode
+import com.qualcomm.robotcore.eventloop.opmode.TeleOp
 import com.qualcomm.robotcore.hardware.Gamepad
 import com.qualcomm.robotcore.util.ElapsedTime
 import org.firstinspires.ftc.robotcontroller.teamcode.GamepadButtons
@@ -99,4 +100,28 @@ abstract class UnifiedTeleOpKt : LinearOpMode() {
 
         for (mechanism in mechanisms) mechanism.stop()
     }
+}
+
+@TeleOp(name = "Kotlin Blue TeleOp Controls (Field)", group = "CompetitionKt")
+class FieldTeleOpOpModeBlueKt : UnifiedTeleOpKt() {
+    override val orientationMode = HardwareMechanismKt.DriveMode.FIELD
+    override val teamColor = TeamColor.BLUE
+}
+
+@TeleOp(name = "Kotlin Red TeleOp Controls (Field)", group = "CompetitionKt")
+class FieldTeleOpOpModeRedKt : UnifiedTeleOpKt() {
+    override val orientationMode = HardwareMechanismKt.DriveMode.FIELD
+    override val teamColor = TeamColor.RED
+}
+
+@TeleOp(name = "Kotlin Blue TeleOp Controls (Robot)", group = "CompetitionKt")
+class RobotTeleOpOpModeBlueKt : UnifiedTeleOpKt() {
+    override val orientationMode = HardwareMechanismKt.DriveMode.ROBOT
+    override val teamColor = TeamColor.BLUE
+}
+
+@TeleOp(name = "Kotlin Red TeleOp Controls (Robot)", group = "CompetitionKt")
+class RobotTeleOpOpModeRedKt : UnifiedTeleOpKt() {
+    override val orientationMode = HardwareMechanismKt.DriveMode.ROBOT
+    override val teamColor = TeamColor.RED
 }
