--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/parameterOpModes/BlueFarAutonomousKt.kt
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.opModes.parameterOpModes
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous
-import org.firstinspires.ftc.teamcode.opModes.UnifiedAutonomousKt
-
-@Autonomous(name = "Kotlin Blue Far Autonomous", group = "CompetitionKt")
-class BlueFarAutonomousKt : UnifiedAutonomousKt() {
-    override val currentLocation = Locations.BlueFar
-}
