--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/parameterOpModes/RedCloseAutonomousKt.kt
@@ -1,9 +1,0 @@
-package org.firstinspires.ftc.teamcode.opModes.parameterOpModes
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous
-import org.firstinspires.ftc.teamcode.opModes.UnifiedAutonomousKt
-
-@Autonomous(name = "Kotlin Red Close Autonomous", group = "CompetitionKt")
-class RedCloseAutonomousKt : UnifiedAutonomousKt() {
-    override val currentLocation = Locations.RedClose
-}
