--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opModes/parameterOpModes/RobotTeleOpOpModeRedKt.kt
@@ -1,12 +1,0 @@
-package org.firstinspires.ftc.teamcode.opModes.parameterOpModes
-
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp
-import org.firstinspires.ftc.robotcontroller.teamcode.HardwareMechanismKt
-import org.firstinspires.ftc.robotcontroller.teamcode.TeamColor
-import org.firstinspires.ftc.teamcode.opModes.UnifiedTeleOpKt
-
-@TeleOp(name = "Kotlin Red TeleOp Controls (Robot)", group = "CompetitionKt")
-class RobotTeleOpOpModeRedKt : UnifiedTeleOpKt() {
-    override val orientationMode = HardwareMechanismKt.DriveMode.ROBOT
-    override val teamColor = TeamColor.RED
-}
